create 'EmptySpaceToken' as base for all spacing tokens, split front matter record parsers to separate files
