+++ src/vs/editor/common/codecs/frontMatterCodec/constants.ts
@@ -5,12 +5,12 @@
 
 import { NewLine } from '../linesCodec/tokens/newLine.js';
 import { CarriageReturn } from '../linesCodec/tokens/carriageReturn.js';
-import { FormFeed, Space, Tab, VerticalTab } from '../simpleCodec/tokens/index.js';
+import { FormFeed, EmptySpaceToken } from '../simpleCodec/tokens/index.js';
 
 /**
- * List of valid "space" tokens that are valid between
- * different records of the Front Matter header.
+ * List of valid "space" tokens that are valid between different
+ * records of a Front Matter header.
  */
-export const VALID_SPACE_TOKENS = Object.freeze([
-	Space, Tab, CarriageReturn, NewLine, FormFeed, VerticalTab,
+export const VALID_INTER_RECORD_SPACING_TOKENS = Object.freeze([
+	EmptySpaceToken, CarriageReturn, NewLine, FormFeed,
 ]);
