+++ src/vs/editor/common/codecs/frontMatterCodec/parsers/frontMatterRecord/index.ts
@@ -1,0 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export { PartialFrontMatterRecord } from './frontMatterRecord.js';
+export { PartialFrontMatterRecordName } from './frontMatterRecordName.js';
+export { PartialFrontMatterRecordNameWithDelimiter } from './frontMatterRecordNameWithDelimiter.js';
