+++ src/vs/editor/common/codecs/simpleCodec/tokens/index.ts
@@ -16,9 +16,9 @@ export { Comma } from './comma.js';
 export { FormFeed } from './formFeed.js';
 export { DollarSign } from './dollarSign.js';
 export { VerticalTab } from './verticalTab.js';
-export { SimpleToken } from './simpleToken.js';
 export { DoubleQuote } from './doubleQuote.js';
 export { ExclamationMark } from './exclamationMark.js';
+export { SimpleToken, EmptySpaceToken } from './simpleToken.js';
 export { type TBracket, LeftBracket, RightBracket } from './brackets.js';
 export { type TCurlyBrace, LeftCurlyBrace, RightCurlyBrace } from './curlyBraces.js';
 export { type TParenthesis, LeftParenthesis, RightParenthesis } from './parentheses.js';
