+++ src/vs/editor/common/codecs/simpleCodec/tokens/simpleToken.ts
@@ -51,3 +51,8 @@ export abstract class SimpleToken<TSymbol extends string> extends BaseToken<TSym
 		));
 	}
 }
+
+/**
+ * TODO: @legomushroom
+ */
+export abstract class EmptySpaceToken extends SimpleToken { }
